+++ CHANGELOG.md
@@ -10,3 +10,13 @@ All notable changes to this project will be documented in this file.
 - Changed: package.json
 - Changed: src/
 - Changed: tsconfig.lib.json
+
+## [2.0.1] - 2025-04-29
+
+- Modified: package.json
+- Modified: src/restful/base.restful.ts
+- Modified: src/restful/hts-hashgraph.restful.ts
+- Modified: src/restful/transactions-hashgraph.restful.ts
+- Modified: src/restful/validators.restful.ts
+- Modified: src/smartnode-sdk.service.ts
+- Modified: src/types/package.json